+++ doc/changelog.rst
@@ -2,6 +2,14 @@
 What's New
 ==========
 
+0.1004.2 (2015-01-03)
+=====================
+
+Hotfix release.
+
+* Fixed PhantomJS mode's MITM proxy AttributeError on certificates.
+
+
 0.1004.1 (2015-01-03)
 =====================